Pass account and timelimit to the new SLURM launchers.
Some SLURM setups require an account to be passed in the user
could possibly be using different accounts. This lets the
user either a) specify an account and timelimit or b)
have a default account and timelimit set by inspecting
which accounts the user has access to that can submit to
the requested queue, and then using a random one of those.

Also set the memory per core to 250M for SLURM launchers
by default, since some setups will give you almost no memory
unless you ask for it.