Respect LSF_UNIT_FOR_LIMITS.
LSF allows you to set the unit for resource limits and we were
forcing KB to be used. If LSF_UNIT_FOR_LIMITS was set to MB then
that lead to behavior where we were requesting 1024 more memory
than we meant to.

This fixes that and defaults to KB if LSF_UNIT_FOR_LIMITS is not
found. Thanks to @ink1 for the report and for digging deep into
figuring out what was happening.