Fix for asking for too much memory on LSF.
If we wanted 8 GB per engine, 1 engine on the machine with
4 cores for each engine, and the resources are per slot (core)
we should be asking for 8 GB/4 cores = 2 GB/core per engine.

If we are jamming 10 engines on the machine and we want 8 GB
per engine and 4 cores for each engine and resources are per
slot (core) then we need a total of 80 GB on the machine,
and 40 cores so we should be asking for 2 GB/core.

Before mem/self.cores for the 10 engine instance was asking
for 8gb/4 cores = 2 gb/core. But then we were multiplying it
by the number of engines, so we were asking for 20 gb/core.