Fix PBSPro support.
PBSPro cannot schedule uninterruptable array jobs, so instead of running the
engines as an array, we create a single engine job submission script and submit
it repeatedly for each job.

This commit also removes some dead code for dill support.

Thanks to Endre Sebestyen (@razZ0r), Francesco Ferrari and Tiziana Castrignanò
for raising the issue, providing an account with access to PBSPro on the Cineca
cluster and testing that the fix works.